requirements.txt
telethon
python-dotenv
flask
requests